Refactor: `WriteDiary` 레이아웃 단순화
- DiaryContent 컴포넌트를 삭제하고 WriteDiary로 통합
- UI 배치 간결화
